+++ lib/src/main/kotlin/cz/eman/swagger/codegen/language/CodegenConst.kt
@@ -18,10 +18,16 @@ const val GENERATE_PRIMITIVE_TYPE_ALIAS = "generatePrimitiveTypeAlias"
 const val GENERATE_PRIMITIVE_TYPE_ALIAS_DESCRIPTION = "Option to generate typealias for primitives."
 
 const val COMPOSED_VARS_NOT_REQUIRED = "composedVarsNotRequired"
-const val COMPOSED_VARS_NOT_REQUIRED_DESCRIPTION = "Option to force variables of ComposedSchema (oneOf, anyOf) to not be required. Default is false."
+const val COMPOSED_VARS_NOT_REQUIRED_DESCRIPTION =
+    "Option to force variables of ComposedSchema (oneOf, anyOf) to not be required. Default is false."
 
 const val REMOVE_OPERATION_PARAMS = "removeOperationParams"
-const val REMOVE_OPERATION_PARAMS_DESCRIPTION = "Option to remove specific parameters from operation. Uses base name to filter the parameters."
+const val REMOVE_OPERATION_PARAMS_DESCRIPTION =
+    "Option to remove specific parameters from operation. Uses base name to filter the parameters."
+
+const val ARRAY_AS_ARRAY_LIST = "arrayAsArrayList"
+const val ARRAY_AS_ARRAY_LIST_DESCRIPTION =
+    "Option to represent Array as ArrayList which can be useful with complex schemas. Use with caution."
 
 const val HEADER_CLI = "headerCliOptions"
 const val HEADER_CLI_DESCRIPTION = "Options to generate Header"
@@ -29,4 +35,5 @@ const val REMOVE_MINUS_TEXT_FROM_HEADER = "removeMinusTextInHeaderProperty"
 const val REMOVE_MINUS_TEXT_FROM_HEADER_DESCRIPTION = "Remove minus text from header's property name if it is present."
 
 const val REMOVE_ENDPOINT_STARTING_SLASH = "ignoreEndpointStartingSlash"
-const val REMOVE_ENDPOINT_STARTING_SLASH_DESCRIPTION = "Remove/Ignore a starting slash from an endpoint definition if it is present."
+const val REMOVE_ENDPOINT_STARTING_SLASH_DESCRIPTION =
+    "Remove/Ignore a starting slash from an endpoint definition if it is present."